requirements.txt
Flask==3.1.1
gunicorn==20.1.0
python-dotenv==1.1.1
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.4
google-cloud-translate==3.15.3
google-cloud-language==2.13.0
PyMySQL==1.1.0
Flask-SQLAlchemy==3.0.5
cryptography==41.0.3
mysql-connector-python==8.2.0
moviepy==1.0.3
opencv-python-headless==4.10.0.84
numpy>=1.21.2
spanlp>=1.1.0
